requirements.txt
numpy
pandas
matplotlib
seaborn
jupyter
python-dotenv
requests
loguru
streamlit
plotly
openai-agents
openai
eval_type_backport
